requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-money==1.1
django-filter==2.3.0
django-cors-headers==3.4.0
django-rq==2.3.2
django-health-check==3.12.1
gunicorn==20.0.4
psycopg2==2.8.5
rq==1.5.0
